Remove all custom add-ons related code
We hid support for installing add-ons months ago, recent updates to
mozilla libraries is making this code more difficult to maintain. Since
we aren't even using it, remove it for now. If we decide to bring back
custom add-on support, we could revert this commit, or re-implement the
feature with the latest APIs.