Re-initialize metadata from play store listing
This backs up the current play store listing as of April 9, 2025, in
case we need to bring back any of the translation.
Reinitialization was done by fully deleting fastlane/metadata/android
and running `bundle fastlane supply init`. Some extra files we add that
aren't available in play store were deleted by doing this, but the
important part is that we have the translated descriptions.